chore: small code style fixes
