+++ geomas/core/rag_modules/rag_pipeline.py
@@ -4,15 +4,14 @@ import importlib.util
 import logging
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Dict, Mapping, Optional, Sequence
-
-import requests
+from typing import Any, Dict, Mapping, Optional, Sequence
 
 from langchain_community.embeddings.sentence_transformer import (
     SentenceTransformerEmbeddings,
 )
 from langchain_core.documents import Document
 
+from geomas.core.inference.lmstudio_client import LmStudioClient
 from geomas.core.inference.ollama_client import OllamaClient
 from geomas.core.rag_modules.data_adapter import DataLoaderAdapter
 from geomas.core.rag_modules.database.chroma_db import (
@@ -80,83 +79,6 @@ def _create_llm_connector(
 
     params = dict(model_params or {})
     return RuntimeLlmConnector(url, params)
-
-
-class LmStudioClient:
-    """Thin HTTP client targeting an LM Studio OpenAI-compatible endpoint."""
-
-    def __init__(
-        self,
-        *,
-        base_url: str,
-        model: str,
-        headers: Mapping[str, str] | None = None,
-        timeout: float | None = None,
-    ) -> None:
-        if not base_url:
-            raise ValueError("LM Studio base_url must be provided")
-        if not model:
-            raise ValueError("LM Studio model must be provided")
-
-        self._endpoint = f"{base_url.rstrip('/')}/chat/completions"
-        self._model = model
-        self._headers = {str(key): str(value) for key, value in dict(headers or {}).items()}
-        self._timeout = timeout
-
-    def generate(
-        self,
-        messages: Sequence[Mapping[str, str]],
-        *,
-        temperature: float,
-    ) -> str:
-        """Send ``messages`` to the LM Studio endpoint and return the response."""
-
-        payload = {
-            "model": self._model,
-            "messages": [dict(message) for message in messages],
-            "temperature": temperature,
-        }
-
-        try:
-            response = requests.post(
-                self._endpoint,
-                json=payload,
-                headers=self._headers or None,
-                timeout=self._timeout,
-            )
-        except requests.RequestException as exc:
-            raise RuntimeError(f"Failed to reach LM Studio endpoint: {exc}") from exc
-
-        if response.status_code >= 400:
-            snippet = response.text.strip()
-            raise RuntimeError(
-                "LM Studio request failed with status "
-                f"{response.status_code}: {snippet or response.reason}"
-            )
-
-        try:
-            data = response.json()
-        except ValueError as exc:
-            raise RuntimeError("LM Studio response was not valid JSON") from exc
-
-        choices = data.get("choices")
-        if not isinstance(choices, Sequence) or not choices:
-            raise RuntimeError("LM Studio response did not include any choices")
-
-        first_choice = choices[0]
-        if isinstance(first_choice, Mapping):
-            message_payload = first_choice.get("message")
-            if isinstance(message_payload, Mapping):
-                content = message_payload.get("content")
-            else:
-                content = first_choice.get("text")
-        else:
-            content = None
-
-        if not content:
-            raise RuntimeError("LM Studio response was missing completion content")
-
-        return str(content)
 
 
 class BaseRAGPipeline(ABC):
